chore(team): split mailbox channel paths
